(PC-18333) feat(shareApp): Open native share modal
