+++ src/features/shareApp/components/ShareAppModalNew.tsx
@@ -1,6 +1,7 @@
 import React, { FunctionComponent, useCallback } from 'react'
 import styled from 'styled-components/native'
 
+import { shareApp } from 'features/shareApp/helpers/shareApp'
 import {
   shareAppModalInformations,
   ShareAppModalType,
@@ -21,6 +22,7 @@ export const ShareAppModalNew: FunctionComponent<Props> = ({ visible, hideModal,
   const openShareAppModal = useCallback(() => {
     analytics.logShareApp(modalType)
     hideModal()
+    setTimeout(shareApp, 0)
   }, [modalType, hideModal])
 
   const onCloseIconPress = useCallback(() => {
