+++ src/libs/share/share.ts
@@ -10,13 +10,17 @@ export type ShareContent = {
 
 type ShareOptions = Parameters<typeof Share.share>[1]
 
-export async function share(content: ShareContent, options: ShareOptions): Promise<void> {
+export async function share(
+  content: ShareContent,
+  options: ShareOptions,
+  logAnalytics?: boolean
+): Promise<void> {
   if (Platform.OS === 'ios' || Platform.OS === 'android') {
     // The only place where it's ok to deactivate the eslint rule is here.
     // Never use Share.share() outside this module.
     // eslint-disable-next-line no-restricted-properties
     const shareAction = await Share.share(content, options)
-    logShareAnalytics(shareAction)
+    if (logAnalytics) logShareAnalytics(shareAction)
   }
   // On web, the share feature is supported by the WebShareModal component.
 }
