--- src/libs/share/shareApp/shareApp.ts
@@ -1,9 +1,0 @@
-import { share } from 'libs/share'
-import { shareAppContent } from 'libs/share/shareApp/shareAppContent'
-
-const shareOptions = {
-  subject: shareAppContent.title, // iOS only
-  dialogTitle: shareAppContent.title, // android only
-}
-
-export const shareApp = () => share(shareAppContent, shareOptions)
